# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend Laravel Api documentation",
        "description": "Backend Laravel API description",
        "contact": {
            "email": "info@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.5.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Api Server"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Perform Login",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "info@example.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "password123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucessful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "The token",
                                            "type": "string",
                                            "example": "ey123456789.a1b2c3d4f5"
                                        },
                                        "token_type": {
                                            "description": "Token Type",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Expiration token (in seconds)",
                                            "type": "number",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "The error",
                                            "type": "string",
                                            "example": "invalid username or password"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Perform token refresh",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Previous token",
                                        "type": "string",
                                        "example": "eya1.b2.c3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucessful refresh",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "The token",
                                            "type": "string",
                                            "example": "ey123456789.a1b2c3d4f5"
                                        },
                                        "token_type": {
                                            "description": "Token Type",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Expiration token (in seconds)",
                                            "type": "number",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "The error",
                                            "type": "string",
                                            "example": "Token could not be parsed from the request."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ping": {
            "get": {
                "tags": [
                    "Ping"
                ],
                "summary": "Return current status of app",
                "responses": {
                    "200": {
                        "description": "Current app status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PingSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Return roles list",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/limit_query"
                    },
                    {
                        "$ref": "#/components/parameters/page_query"
                    },
                    {
                        "$ref": "#/components/parameters/sort_query"
                    },
                    {
                        "$ref": "#/components/parameters/direction_sort"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolesSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create new role",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "role": {
                                        "description": "The name of the new role",
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succcess": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error on body"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Return users list",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/limit_query"
                    },
                    {
                        "$ref": "#/components/parameters/page_query"
                    },
                    {
                        "$ref": "#/components/parameters/sort_query"
                    },
                    {
                        "$ref": "#/components/parameters/direction_sort"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PingSchema": {
                "title": "Ping",
                "description": "Ping model",
                "properties": {
                    "data": {
                        "description": "Ping status",
                        "type": "string",
                        "items": {
                            "properties": {
                                "app_version": {
                                    "description": "Backend current version",
                                    "type": "string",
                                    "example": "1.0.0"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "RolesSchema": {
                "title": "Roles",
                "description": "Roles model",
                "properties": {
                    "data": {
                        "description": "List of roles",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID of the role",
                                    "type": "number",
                                    "example": 1
                                },
                                "role": {
                                    "description": "Name of the role",
                                    "type": "string",
                                    "example": "admin"
                                },
                                "users": {
                                    "description": "List of associated Users",
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the user",
                                                "type": "number",
                                                "example": 1
                                            },
                                            "email": {
                                                "description": "Email",
                                                "type": "string",
                                                "example": "info@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "RolesPostSchema": {
                "title": "Roles",
                "description": "Roles model",
                "properties": {
                    "data": {
                        "description": "New Role",
                        "type": "object",
                        "items": {
                            "properties": {
                                "role": {
                                    "description": "Name of the role",
                                    "type": "string",
                                    "example": "admin"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UsersSchema": {
                "title": "Users",
                "description": "Users model",
                "properties": {
                    "data": {
                        "description": "List of users",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID of the user",
                                    "type": "number",
                                    "example": 1
                                },
                                "email": {
                                    "description": "Email",
                                    "type": "string",
                                    "example": "info@example.com"
                                },
                                "roles": {
                                    "description": "Associated roles",
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the role",
                                                "type": "number",
                                                "example": 1
                                            },
                                            "role": {
                                                "description": "Role name",
                                                "type": "string",
                                                "example": "admin"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "page_query": {
                "name": "page",
                "in": "query",
                "description": "Page number of results",
                "schema": {
                    "type": "number",
                    "default": 1
                }
            },
            "limit_query": {
                "name": "per_page",
                "in": "query",
                "description": "Limit the number of results",
                "schema": {
                    "type": "number",
                    "default": 15
                }
            },
            "sort_query": {
                "name": "sort",
                "in": "query",
                "description": "Sort the result",
                "schema": {
                    "type": "string",
                    "default": "id"
                }
            },
            "direction_sort": {
                "name": "direction",
                "in": "query",
                "description": "Direction of sort",
                "schema": {
                    "type": "string",
                    "default": "asc"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}